Combine tests related to QueryValues
